refactor(settings): rename browser policy section
