+++ apps/desktop/src/stores/setting/sections/browserPolicy.ts
